move Context into Core
